api/requirements.txt
Django==3.0.6
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
pytest-django==3.9.0
pytest==5.4.2